requirements.txt
Flask
Flask-Mail
python-dotenv
gunicorn app:app
